refactor: Changed spinner css to match figma design. Move Spinner to separate folder with separate styles to improve readability and code structure. Update Spinner import paths across multiple components because of moving spinner to separate folder
