+++ client/src/admin/components/ProductManagement/ProductForm/CategorySelect.jsx
@@ -1,6 +1,6 @@
 // import cl from './index.module.scss'
 
-import Spinner from '@components/helpers/Spinner';
+import Spinner from '@components/helpers/Spinner/Spinner';
 import { loadCategories, selectCategories } from '@redux/selectors';
 import { Controller } from 'react-hook-form';
 import { useSelector } from 'react-redux';
