+++ client/src/admin/components/ProductManagement/ProductList/ProductList.jsx
@@ -1,4 +1,4 @@
-import Spinner from '@components/helpers/Spinner';
+import Spinner from '@components/helpers/Spinner/Spinner';
 import cl from './index.module.scss';
 import { loadProducts, selectProducts } from '@redux/selectors';
 import { useDispatch, useSelector } from 'react-redux';
