+++ client/src/components/ProductPage/ProductSet.jsx
@@ -2,7 +2,7 @@ import SliderBoxMain from '@components/UI/Sliders/SliderBoxMain';
 import Heading from '@components/UI/Texts/Heading/Heading';
 import cl from './index.module.scss';
 import { useSelector } from 'react-redux';
-import Spinner from '@components/helpers/Spinner';
+import Spinner from '@components/helpers/Spinner/Spinner';
 import { loadSets, selectSets } from '@redux/selectors';
 import useTranslationNamespace from '@hooks/useTranslationNamespace';
