+++ client/src/components/helpers/Loading.jsx
@@ -4,7 +4,6 @@ export default function Loading() {
   return (
     <div className={cl.wrapper}>
       <div className={cl.cloud}></div>
-      {/* <div className={cl.dotts}></div> */}
     </div>
   );
 }
