+++ client/src/components/helpers/Spinner/index.module.scss
@@ -1,0 +1,28 @@
+@import '@mixins';
+
+.wrapSpinner {
+  @include flex(row, center, center);
+  box-sizing: border-box;
+  width: 100%;
+  height: 100%;
+
+  .spinner {
+    width: 48px;
+    height: 48px;
+    border: 5px solid #c0c0c0;
+    border-bottom-color: transparent;
+    border-radius: 50%;
+    display: inline-block;
+    box-sizing: border-box;
+    animation: rotation 1s linear infinite;
+  }
+
+  @keyframes rotation {
+    0% {
+      transform: rotate(0deg);
+    }
+    100% {
+      transform: rotate(360deg);
+    }
+  }
+}
