+++ client/src/components/helpers/index.module.scss
@@ -103,149 +103,3 @@
     transform: rotate(-360deg);
   }
 }
-
-// *** LOADER DOTS ***
-.dotts {
-  color: #fff;
-  font-size: 10px;
-  width: 1em;
-  height: 1em;
-  border-radius: 50%;
-  position: relative;
-  text-indent: -9999em;
-  animation: mulShdSpin 1.3s infinite linear;
-  transform: translateZ(0);
-}
-
-@keyframes mulShdSpin {
-  0%,
-  100% {
-    box-shadow:
-      0 -3em 0 0.2em,
-      2em -2em 0 0em,
-      3em 0 0 -1em,
-      2em 2em 0 -1em,
-      0 3em 0 -1em,
-      -2em 2em 0 -1em,
-      -3em 0 0 -1em,
-      -2em -2em 0 0;
-  }
-  12.5% {
-    box-shadow:
-      0 -3em 0 0,
-      2em -2em 0 0.2em,
-      3em 0 0 0,
-      2em 2em 0 -1em,
-      0 3em 0 -1em,
-      -2em 2em 0 -1em,
-      -3em 0 0 -1em,
-      -2em -2em 0 -1em;
-  }
-  25% {
-    box-shadow:
-      0 -3em 0 -0.5em,
-      2em -2em 0 0,
-      3em 0 0 0.2em,
-      2em 2em 0 0,
-      0 3em 0 -1em,
-      -2em 2em 0 -1em,
-      -3em 0 0 -1em,
-      -2em -2em 0 -1em;
-  }
-  37.5% {
-    box-shadow:
-      0 -3em 0 -1em,
-      2em -2em 0 -1em,
-      3em 0em 0 0,
-      2em 2em 0 0.2em,
-      0 3em 0 0em,
-      -2em 2em 0 -1em,
-      -3em 0em 0 -1em,
-      -2em -2em 0 -1em;
-  }
-  50% {
-    box-shadow:
-      0 -3em 0 -1em,
-      2em -2em 0 -1em,
-      3em 0 0 -1em,
-      2em 2em 0 0em,
-      0 3em 0 0.2em,
-      -2em 2em 0 0,
-      -3em 0em 0 -1em,
-      -2em -2em 0 -1em;
-  }
-  62.5% {
-    box-shadow:
-      0 -3em 0 -1em,
-      2em -2em 0 -1em,
-      3em 0 0 -1em,
-      2em 2em 0 -1em,
-      0 3em 0 0,
-      -2em 2em 0 0.2em,
-      -3em 0 0 0,
-      -2em -2em 0 -1em;
-  }
-  75% {
-    box-shadow:
-      0em -3em 0 -1em,
-      2em -2em 0 -1em,
-      3em 0em 0 -1em,
-      2em 2em 0 -1em,
-      0 3em 0 -1em,
-      -2em 2em 0 0,
-      -3em 0em 0 0.2em,
-      -2em -2em 0 0;
-  }
-  87.5% {
-    box-shadow:
-      0em -3em 0 0,
-      2em -2em 0 -1em,
-      3em 0 0 -1em,
-      2em 2em 0 -1em,
-      0 3em 0 -1em,
-      -2em 2em 0 0,
-      -3em 0em 0 0,
-      -2em -2em 0 0.2em;
-  }
-}
-
-// *** SPINNER ***
-.wrapSpinner {
-  @include flex(row, center, center);
-  box-sizing: border-box;
-  width: 100%;
-  height: 100%;
-
-  .spinner {
-    width: 16px;
-    height: 16px;
-    border-radius: 50%;
-    background-color: #8f8f8f;
-    box-shadow:
-      32px 0 #8f8f8f,
-      -32px 0 #8f8f8f;
-    position: relative;
-    animation: flash 0.5s ease-out infinite alternate;
-  }
-
-  @keyframes flash {
-    0% {
-      background-color: #fff2;
-      box-shadow:
-        32px 0 #fff2,
-        -32px 0 #8f8f8f;
-    }
-    50% {
-      background-color: #8f8f8f;
-      box-shadow:
-        32px 0 #fff2,
-        -32px 0 #fff2;
-    }
-    100% {
-      background-color: #fff2;
-      box-shadow:
-        32px 0 #8f8f8f,
-        -32px 0 #fff2;
-    }
-  }
-}
