+++ client/src/pages/CatalogPage/Aside/FilterType/FilterByType.jsx
@@ -4,7 +4,7 @@ import useTranslationNamespace from '@hooks/useTranslationNamespace';
 import Heading from '@UI/Texts/Heading/Heading';
 import { useDispatch, useSelector } from 'react-redux';
 import EmptyPage from '@components/helpers/EmptyPage';
-import Spinner from '@components/helpers/Spinner';
+import Spinner from '@components/helpers/Spinner/Spinner';
 import { filterType, loadTypes, selectTypes } from '@redux/selectors';
 import { fetchProductsAll, getProductsByTypes } from '@redux/products/service';
 import { setType } from '@redux/filter/filterSlice';
