+++ client/src/pages/CatalogPage/FilterCategory/FilterByCategory.jsx
@@ -1,7 +1,7 @@
 import cl from './index.module.scss';
 import { useSelector } from 'react-redux';
 import EmptyPage from '@components/helpers/EmptyPage';
-import Spinner from '@components/helpers/Spinner';
+import Spinner from '@components/helpers/Spinner/Spinner';
 import { loadCategories, selectCategories } from '@redux/selectors';
 
 import useScreenSizes from '@hooks/useScreenSizes';
