+++ client/src/pages/HomePage/MainSliders/BestSellers.jsx
@@ -5,7 +5,7 @@ import SliderBoxMain from '@components/UI/Sliders/SliderBoxMain';
 import { useSelector } from 'react-redux';
 import EmptyPage from '@components/helpers/EmptyPage';
 import { loadBestSeller, selectBestSeller } from '@redux/selectors';
-import Spinner from '@components/helpers/Spinner';
+import Spinner from '@components/helpers/Spinner/Spinner';
 
 const BestSellers = () => {
   const { getTranslation } = useTranslationNamespace('common');
